+++ User/decision.h
@@ -20,8 +20,8 @@ struct board_derived_info_t
 	std::vector<player_t*> villains_after_hero;
 	player_t* main_villain;
 	player_t* secondary_villain;
-	double pot;
-	double current_bet;
+	float pot;
+	float current_bet;
 
 	
 	// NOTE: this can be expanded nicely 
@@ -31,7 +31,7 @@ struct board_derived_info_t
 struct decision_t
 {
 	hand_action_t action;
-	double sum;
+	float sum;
 };
