+++ User/player.h
@@ -22,9 +22,9 @@ enum board_stage_t
 struct board_stage_stats_t
 {
 	count_t seen;
-	double aggression;
-	double aggression_freq;
-	double avg_hand_strength; // the average strength of the hand that reached this stage
+	float aggression;
+	float aggression_freq;
+	float avg_hand_strength; // the average strength of the hand that reached this stage
 	count_t win;
 	count_t lose;
 
@@ -43,9 +43,9 @@ struct board_stage_stats_t
 	count_t raised_cbet;
 
 	// weighted in accordance pot size
-	double weighted_avg_call;
-	double weighted_avg_bet;
-	double weighted_avg_raise;
+	float weighted_avg_call;
+	float weighted_avg_bet;
+	float weighted_avg_raise;
 
 	count_t dealer_raise;
 	count_t small_raise;
@@ -57,12 +57,12 @@ struct player_t
 {
 	std::string label;  // this label might be p6, for example, but the player might be Player 1
 	std::string name;
-	double balance = 0;
-	double current_bet = 0;
-	double overall_agression = 0;
-	double overall_agression_freq = 0;
+	float balance = 0;
+	float current_bet = 0;
+	float overall_agression = 0;
+	float overall_agression_freq = 0;
 	count_t hand_seen = 0;
-	double overall_avg_hand_strength = 0;
+	float overall_avg_hand_strength = 0;
 	count_t win = 0;
 	count_t lose = 0;
 	count_t faceup_lose = 0;
@@ -74,7 +74,7 @@ struct player_t
 	bool is_big = false;
 	bool is_in_hand = false;
 	bool is_in_game = false;  // used for folds/sitouts
-	double range[PLAYERS_COUNT] = { 0 };  // the weakest cards this player plays in a position 
+	float range[PLAYERS_COUNT] = { 0 };  // the weakest cards this player plays in a position 
 
 	board_stage_stats_t preflop = { 0 };
 	board_stage_stats_t flop = { 0 };
