+++ UserTests/test.cpp
@@ -18,10 +18,10 @@ int main(int argc, char** argv)
 	hand_t* hero_hand = get_hand(get_card(_A, H), get_card(_A, C));
 	hand_t* villain_hand = get_hand(get_card(_T, H), get_card(_6, C));
 	std::cout << calc_prwin_vs_hand(hero_hand, villain_hand, &board) << '\n';
-	double start = clock();
+	float start = clock();
 	std::cout << calc_prwin_vs_any_hand(hero_hand, &board) << '\n';
-	double end = clock();
-	double time = (end - start) / CLOCKS_PER_SEC;
+	float end = clock();
+	float time = (end - start) / CLOCKS_PER_SEC;
 	std::cout << time << '\n';
 	return RUN_ALL_TESTS();
 }
